Extract the flushing logic in the tracer
